+++ predictables/corr/src/__init__.py
@@ -1,7 +1,7 @@
-from .bin_bin import calc_binary_binary_corr  # noqa F401
-from .bin_cat import calc_binary_categorical_corr  # noqa F401
-from .cat_cat import calc_categorical_categorical_corr  # noqa F401
-from .cts_bin import calc_continuous_binary_corr  # noqa F401
-from .cts_cat import calc_continuous_categorical_corr  # noqa F401
-from .cts_cts import calc_continuous_continuous_corr  # noqa F401
-from .predictor_target_corr import predictor_target_corr  # noqa F401
+from ._bin_bin import calc_binary_binary_corr  # noqa F401
+from ._bin_cat import calc_binary_categorical_corr  # noqa F401
+from ._cat_cat import calc_categorical_categorical_corr  # noqa F401
+from ._cts_bin import calc_continuous_binary_corr  # noqa F401
+from ._cts_cat import calc_continuous_categorical_corr  # noqa F401
+from ._cts_cts import calc_continuous_continuous_corr  # noqa F401
+from ._predictor_target_corr import predictor_target_corr  # noqa F401
