+++ predictables/corr/src/_predictor_target_corr.py
@@ -3,12 +3,12 @@ from tqdm import tqdm
 
 from predictables.util import get_column_dtype, to_pd_df, to_pd_s
 
-from .bin_bin import calc_binary_binary_corr
-from .bin_cat import calc_binary_categorical_corr
-from .cat_cat import calc_categorical_categorical_corr
-from .cts_bin import calc_continuous_binary_corr
-from .cts_cat import calc_continuous_categorical_corr
-from .cts_cts import calc_continuous_continuous_corr
+from ._bin_bin import calc_binary_binary_corr
+from ._bin_cat import calc_binary_categorical_corr
+from ._cat_cat import calc_categorical_categorical_corr
+from ._cts_bin import calc_continuous_binary_corr
+from ._cts_cat import calc_continuous_categorical_corr
+from ._cts_cts import calc_continuous_continuous_corr
 
 
 def predictor_target_corr(X, y):
