+++ predictables/feature_importance/src/forward_stepwise/__init__.py
@@ -1,1 +1,1 @@
-from .evaluate_best_feature_to_add import evaluate_best_feature_to_add  # noqa F401
+from ._evaluate_best_feature_to_add import evaluate_best_feature_to_add  # noqa F401
