+++ predictables/feature_selection/forward_stepwise/src/__init__.py
@@ -1,4 +1,8 @@
-from src.feature_importance import feature_importance  # noqa F401
-from src.fit_model import fit_model  # noqa F401
-from src.plot import plot  # noqa F401
-from src.select_next_feature import select_next_feature  # noqa F401
+from predictables.feature_selection.forward_stepwise.src._feature_importance import (
+    feature_importance,
+)  # noqa F401
+from predictables.feature_selection.forward_stepwise.src._fit_model import fit_model  # noqa F401
+from predictables.feature_selection.forward_stepwise.src._plot import plot  # noqa F401
+from predictables.feature_selection.forward_stepwise.src._select_next_feature import (
+    select_next_feature,
+)  # noqa F401
