+++ predictables/lookups/src/__init__.py
@@ -1,1 +1,1 @@
-from .generate_date_lookup import generate_date_lookup  # noqa F401
+from ._generate_date_lookup import generate_date_lookup  # noqa F401
