+++ predictables/model_analysis/src/__init__.py
@@ -1,2 +1,2 @@
-from .calculate_error_metrics import calculate_error_metrics  # noqa F401
-from .validate_input import validate_input  # noqa F401
+from ._calculate_error_metrics import calculate_error_metrics  # noqa F401
+from ._validate_input import validate_input  # noqa F401
