+++ predictables/pca/src/__init__.py
@@ -1,10 +1,10 @@
-from .bootstrap_pca import bootstrap_pca  # noqa F401
-from .create_biplot import create_biplot  # noqa F401
-from .create_loading_plot import create_loading_plot  # noqa F401
-from .create_scree_plot import create_scree_plot  # noqa F401
-from .feature_importance import (  # noqa F401
+from ._bootstrap_pca import bootstrap_pca  # noqa F401
+from ._create_biplot import create_biplot  # noqa F401
+from ._create_loading_plot import create_loading_plot  # noqa F401
+from ._create_scree_plot import create_scree_plot  # noqa F401
+from ._feature_importance import (  # noqa F401
     pca_feature_importance as feature_importance,
 )
-from .perform_pca import perform_pca  # noqa F401
-from .preprocessing import preprocess_data_for_pca  # noqa F401
-from .select_principal_components import select_n_components_for_variance  # noqa F401
+from ._perform_pca import perform_pca  # noqa F401
+from ._preprocessing import preprocess_data_for_pca  # noqa F401
+from ._select_principal_components import select_n_components_for_variance  # noqa F401
