+++ predictables/pca/src/_perform_pca.py
@@ -6,7 +6,7 @@ from sklearn.decomposition import PCA
 
 from predictables.util import to_pd_df
 
-from .preprocessing import preprocess_data_for_pca
+from ._preprocessing import preprocess_data_for_pca
 
 
 def perform_pca(
