+++ predictables/pca/src/tests/test_feature_importance.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from ..feature_importance import pca_feature_importance
-from ..perform_pca import perform_pca
+from .._feature_importance import pca_feature_importance
+from .._perform_pca import perform_pca
 
 
 @pytest.fixture
@@ -37,8 +37,8 @@ def test_no_normalization(pca):
 
 def test_output_shape(pca):
     feature_importances = pca_feature_importance(pca)
-    assert feature_importances.shape == (
-        3,
+    assert (
+        feature_importances.shape == (3,)
     ), f"Feature importances are {feature_importances}. Expected shape (3,), got {feature_importances.shape}"
